cherry-pick #1122

+++ ams/ams-server/src/main/java/com/netease/arctic/ams/server/service/impl/OrphanFilesCleanService.java
@@ -19,19 +19,17 @@
 package com.netease.arctic.ams.server.service.impl;
 
 import com.netease.arctic.ams.api.Constants;
-import com.netease.arctic.ams.api.DataFileInfo;
 import com.netease.arctic.ams.server.model.TableMetadata;
 import com.netease.arctic.ams.server.service.IOrphanFilesCleanService;
 import com.netease.arctic.ams.server.service.ServiceContainer;
 import com.netease.arctic.ams.server.utils.HiveLocationUtils;
 import com.netease.arctic.ams.server.utils.ScheduledTasks;
 import com.netease.arctic.ams.server.utils.ThreadPool;
+import com.netease.arctic.ams.server.utils.UnKeyedTableUtil;
 import com.netease.arctic.catalog.ArcticCatalog;
 import com.netease.arctic.catalog.CatalogLoader;
 import com.netease.arctic.io.ArcticFileIO;
 import com.netease.arctic.table.ArcticTable;
-import com.netease.arctic.table.BaseTable;
-import com.netease.arctic.table.ChangeTable;
 import com.netease.arctic.table.KeyedTable;
 import com.netease.arctic.table.TableIdentifier;
 import com.netease.arctic.table.TableProperties;
@@ -39,11 +37,9 @@ import com.netease.arctic.table.UnkeyedTable;
 import com.netease.arctic.utils.FileUtil;
 import org.apache.hadoop.fs.FileStatus;
 import org.apache.hadoop.fs.Path;
-import org.apache.iceberg.FileScanTask;
 import org.apache.iceberg.ManifestFile;
 import org.apache.iceberg.ReachableFileUtil;
 import org.apache.iceberg.Snapshot;
-import org.apache.iceberg.io.CloseableIterable;
 import org.apache.iceberg.relocated.com.google.common.collect.Iterables;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
@@ -125,61 +121,50 @@ public class OrphanFilesCleanService implements IOrphanFilesCleanService {
 
   public static void clean(ArcticTable arcticTable, long lastTime, boolean execute,
                            String mode, boolean metadata) {
+    Set<String> validFiles = new HashSet<>();
+    // For clean data files, should get valid files in the base store and the change store, so acquire in advance
+    // to prevent repeated acquisition
+    if (!metadata) {
+      validFiles = getValidDataFiles(arcticTable);
+    }
     if (arcticTable.isKeyedTable()) {
       KeyedTable keyedArcticTable = arcticTable.asKeyedTable();
       if (Constants.INNER_TABLE_BASE.equals(mode)) {
-        clearInternalTable(keyedArcticTable, keyedArcticTable.baseTable(), lastTime, execute, metadata);
+        clearInternalTable(keyedArcticTable, keyedArcticTable.baseTable(), lastTime, execute, metadata, validFiles);
       } else if (Constants.INNER_TABLE_CHANGE.equals(mode)) {
         if (keyedArcticTable.primaryKeySpec().primaryKeyExisted()) {
-          clearInternalTable(keyedArcticTable, keyedArcticTable.changeTable(), lastTime, execute, metadata);
+          clearInternalTable(keyedArcticTable, keyedArcticTable.changeTable(), lastTime, execute, metadata, validFiles);
         } else {
           throw new IllegalStateException("no pk table, only support mode=all/base");
         }
       } else if ("all".equals(mode)) {
-        clearInternalTable(keyedArcticTable, keyedArcticTable.baseTable(), lastTime, execute, metadata);
-        clearInternalTable(keyedArcticTable, keyedArcticTable.changeTable(), lastTime, execute, metadata);
+        clearInternalTable(keyedArcticTable, keyedArcticTable.baseTable(), lastTime, execute, metadata, validFiles);
+        clearInternalTable(keyedArcticTable, keyedArcticTable.changeTable(), lastTime, execute, metadata, validFiles);
       } else {
         throw new IllegalStateException("only support mode=all/base/change");
       }
     } else {
-      clearInternalTable(arcticTable, arcticTable.asUnkeyedTable(), lastTime, execute, metadata);
+      clearInternalTable(arcticTable, arcticTable.asUnkeyedTable(), lastTime, execute, metadata, validFiles);
     }
   }
 
   private static void clearInternalTable(ArcticTable table, UnkeyedTable internalTable, long lastTime,
-                                         boolean execute, boolean metadata) {
+                                         boolean execute, boolean metadata, Set<String> exclude) {
     if (metadata) {
       clearInternalTableMetadata(table, internalTable, lastTime, execute);
     } else {
-      clearInternalTableDataFiles(table, internalTable, lastTime, execute);
+      clearInternalTableDataFiles(table, internalTable, lastTime, execute, exclude);
     }
   }
 
   private static void clearInternalTableDataFiles(ArcticTable table, UnkeyedTable internalTable, long lastTime,
-                                                  boolean execute) {
-    Set<String> validFiles = getValidDataFiles(table.id(), table.io(), internalTable);
-    LOG.info("{} table get {} valid files", table.id(), validFiles.size());
+                                                  boolean execute, Set<String> exclude) {
     int deleteFilesCnt = 0;
-    Set<String> exclude = new HashSet<>();
-    if (internalTable instanceof BaseTable) {
-      List<DataFileInfo> dataFilesInfo = ServiceContainer.getFileInfoCacheService()
-          .getOptimizeDatafiles(table.id().buildTableIdentifier(), Constants.INNER_TABLE_CHANGE);
-      exclude = dataFilesInfo.stream().map(DataFileInfo::getPath).collect(Collectors.toSet());
-    } else if (internalTable instanceof ChangeTable) {
-      List<DataFileInfo> dataFilesInfo = ServiceContainer.getFileInfoCacheService()
-          .getOptimizeDatafiles(table.id().buildTableIdentifier(), Constants.INNER_TABLE_BASE);
-      exclude = dataFilesInfo.stream().map(DataFileInfo::getPath).collect(Collectors.toSet());
-    }
-
-    // add hive location to exclude
-    exclude.addAll(HiveLocationUtils.getHiveLocation(table));
-
     String dataLocation = internalTable.location() + File.separator + DATA_FOLDER_NAME;
     if (table.io().exists(dataLocation)) {
       for (FileStatus fileStatus : table.io().list(dataLocation)) {
         deleteFilesCnt += deleteInvalidDataFiles(table.io(),
             fileStatus,
-            validFiles,
             lastTime,
             exclude,
             execute);
@@ -212,7 +197,6 @@ public class OrphanFilesCleanService implements IOrphanFilesCleanService {
 
   private static int deleteInvalidDataFiles(ArcticFileIO io,
                                             FileStatus fileStatus,
-                                            Set<String> validFiles,
                                             Long lastTime,
                                             Set<String> exclude,
                                             boolean execute) {
@@ -222,7 +206,7 @@ public class OrphanFilesCleanService implements IOrphanFilesCleanService {
         LOG.info("start orphan files clean in {}", location);
         int deleteFileCnt = 0;
         for (FileStatus file : io.list(location)) {
-          deleteFileCnt += deleteInvalidDataFiles(io, file, validFiles, lastTime, exclude, execute);
+          deleteFileCnt += deleteInvalidDataFiles(io, file, lastTime, exclude, execute);
         }
         LOG.info("delete[{}] {} files in {}", execute, deleteFileCnt, location);
 
@@ -247,17 +231,15 @@ public class OrphanFilesCleanService implements IOrphanFilesCleanService {
         return 0;
       }
     } else {
-      if (!validFiles.contains(location) &&
+      if (execute &&
+          !exclude.contains(location) &&
+          !exclude.contains(FileUtil.getUriPath(new Path(location).getParent().toString())) &&
           fileStatus.getModificationTime() < lastTime) {
-        if (execute &&
-            !exclude.contains(location) &&
-            !exclude.contains(new Path(location).getParent().toString())) {
-          io.deleteFile(location);
-        }
+        io.deleteFile(location);
         return 1;
-      } else {
-        return 0;
       }
+
+      return 0;
     }
   }
 
@@ -311,38 +293,32 @@ public class OrphanFilesCleanService implements IOrphanFilesCleanService {
 
     return validFiles;
   }
-  
+
   protected static String getUriPath(String path) {
     return URI.create(path).getPath();
   }
 
-  private static Set<String> getValidDataFiles(TableIdentifier tableIdentifier, ArcticFileIO io,
-                                               UnkeyedTable internalTable) {
+  private static Set<String> getValidDataFiles(ArcticTable arcticTable) {
     Set<String> validFiles = new HashSet<>();
-    Iterable<Snapshot> snapshots = internalTable.snapshots();
-    int size = Iterables.size(snapshots);
-    LOG.info("{} get {} snapshots to scan", tableIdentifier, size);
-    int cnt = 0;
-    for (Snapshot snapshot : snapshots) {
-      cnt++;
-      int before = validFiles.size();
-      io.doAs(() -> {
-        // valid data files
-
-        try (CloseableIterable<FileScanTask> fileScanTasks =
-                 internalTable.newScan().useSnapshot(snapshot.snapshotId()).planFiles()) {
-          for (FileScanTask scanTask : fileScanTasks) {
-            if (scanTask.file() != null) {
-              validFiles.add(getUriPath(scanTask.file().path().toString()));
-              scanTask.deletes().forEach(file -> validFiles.add(getUriPath(file.path().toString())));
-            }
-          }
-          return null;
-        }
-      });
-      LOG.info("{} scan snapshot {}: {} and get {} files, complete {}/{}", tableIdentifier, snapshot.snapshotId(),
-          formatTime(snapshot.timestampMillis()), validFiles.size() - before, cnt, size);
+    if (arcticTable.isKeyedTable()) {
+      Set<String> baseValidFiles = UnKeyedTableUtil.getAllContentFilePath(arcticTable.asKeyedTable().baseTable());
+      LOG.info("{} get {} valid files in the base store", arcticTable.id(), baseValidFiles.size());
+      Set<String> changeValidFiles = UnKeyedTableUtil.getAllContentFilePath(arcticTable.asKeyedTable().changeTable());
+      LOG.info("{} get {} valid files in the change store", arcticTable.id(), baseValidFiles.size());
+      validFiles.addAll(baseValidFiles);
+      validFiles.addAll(changeValidFiles);
+    } else {
+      Set<String> baseValidFiles = UnKeyedTableUtil.getAllContentFilePath(arcticTable.asUnkeyedTable());
+      LOG.info("{} get {} valid files in the base store", arcticTable.id(), baseValidFiles.size());
+      validFiles.addAll(baseValidFiles);
     }
+
+    LOG.info("{} get {} valid files", arcticTable.id(), validFiles.size());
+
+    // add hive location to exclude
+    Set<String> hiveValidLocations = HiveLocationUtils.getHiveLocation(arcticTable);
+    LOG.info("{} get {} valid locations in the Hive", arcticTable.id(), hiveValidLocations.size());
+    validFiles.addAll(hiveValidLocations);
 
     return validFiles;
   }
